UHF-89: Moved object initialization to correct file
