+++ src/Plugin/migrate/destination/Unit.php
@@ -4,10 +4,6 @@ declare(strict_types = 1);
 
 namespace Drupal\helfi_tpr\Plugin\migrate\destination;
 
-use Drupal\Core\Language\LanguageManagerInterface;
-use Drupal\migrate\Plugin\MigrationInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
 /**
  * Provides a destination plugin for Tpr entities.
  *
@@ -16,29 +12,6 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * )
  */
 final class Unit extends Tpr {
-
-  /**
-   * The language manager service.
-   *
-   * @var \Drupal\Core\Language\LanguageManagerInterface
-   */
-  protected LanguageManagerInterface $languageManager;
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(
-    ContainerInterface $container,
-    array $configuration,
-    $plugin_id,
-    $plugin_definition,
-    MigrationInterface $migration = NULL
-  ) {
-    $instance = parent::create($container, $configuration, $plugin_id, $plugin_definition, $migration);
-    $instance->languageManager = $container->get('language_manager');
-
-    return $instance;
-  }
 
   /**
    * {@inheritdoc}
